+++ src/main/java/com/example/customerserver/web/controller/AuthController.java
@@ -5,6 +5,7 @@ import com.example.customerserver.web.response.TokenResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
@@ -27,7 +28,8 @@ public class AuthController {
     }
 
     @RequestMapping(path = "/login", method = {RequestMethod.GET, RequestMethod.POST})
-    public String login() {
+    public String login(@RequestParam(defaultValue = "false") boolean error, Model model) {
+        model.addAttribute("error", error);
         return "signin";
     }
