+++ src/main/resources/templates/signin.html
@@ -90,7 +90,7 @@
                                 </div>
                                 <hr>
                                 <div class="text-center">
-                                    <a class="small" th:href="@{/auth/signup}">Create an Account!</a>
+                                    <a class="small" th:href="@{/signup}">Create an Account!</a>
                                 </div>
                             </div>
                         </div>
